Reorganize package to have more nested modules
