+++ pdf_annotate/__init__.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 from pdf_annotate.annotator import PdfAnnotator
-from pdf_annotate.appearance import Appearance
-from pdf_annotate.location import Location
-from pdf_annotate.metadata import Metadata
+from pdf_annotate.config.appearance import Appearance
+from pdf_annotate.config.location import Location
+from pdf_annotate.config.metadata import Metadata
 
 
 __all__ = ['PdfAnnotator', 'Appearance', 'Location', 'Metadata']
